+++ forms/book.py
@@ -15,7 +15,7 @@ from forms.custom_validators import \
     check_pub_date, title_book_exists
 
 
-class BookForm(FlaskForm):
+class LibraryItemForm(FlaskForm):
     languages = [('polish', 'Polish'),
                  ('english', 'English'),
                  ('other', 'Other')]
@@ -24,28 +24,11 @@ class BookForm(FlaskForm):
                   ('magazines', 'Magazines'),
                   ('other', 'Other')]
 
+    # LibraryItem
     radio = RadioField("radio",
                        choices=[('book', 'Book'),
                                 ('magazine', 'Magazine')],
                        render_kw=({'class': 'radio_but'}))
-
-    """LibraryItem"""
-    title = StringField('Title',
-                        validators=[DataRequired(),
-                                    Length(3),
-                                    title_book_exists],
-                        render_kw=({'class': 'inputs',
-                                    'id': 'title',
-                                    'placeholder': 'Title',
-                                    'disabled': True}))
-    title_of_magazine = StringField('Title',
-                                    validators=[DataRequired(),
-                                                Length(3)],
-                                    render_kw=({'class': 'inputs',
-                                                'id': 'title_of_magazine',
-                                                'placeholder':
-                                                    'Title of magazine',
-                                                'disabled': True}))
 
     table_of_contents = TextAreaField('Table of contents',
                                       validators=[DataRequired(),
@@ -87,7 +70,34 @@ class BookForm(FlaskForm):
                                             'placeholder': 'Description',
                                             'disabled': True}))
 
-    """Book"""
+    pub_date = SelectField('Year of publication',
+                           choices=[(str(year), str(year))
+                                    for year in
+                                    range(1970,
+                                          datetime.now().year + 1)],
+                           validators=[check_pub_date],
+                           render_kw=({
+                               'class': 'custom-select mb-2 mr-sm-2 mb-sm-0',
+                               'id': 'mySelect',
+                               'placeholder': 'Year of publication',
+                               'disabled': True}))
+
+    submit = SubmitField('Update',
+                         render_kw=({'class': 'btn btn-primary submits',
+                                     'id': 'button',
+                                     'disabled': True}))
+
+
+class BookForm(LibraryItemForm):
+    title = StringField('Title',
+                        validators=[DataRequired(),
+                                    Length(3),
+                                    title_book_exists],
+                        render_kw=({'class': 'inputs',
+                                    'id': 'title',
+                                    'placeholder': 'Title',
+                                    'disabled': True}))
+
     isbn = StringField('ISBN number',
                        validators=[DataRequired(), check_isbn],
                        render_kw=({'class': 'inputs',
@@ -108,19 +118,7 @@ class BookForm(FlaskForm):
                                         'placeholder': 'Publisher',
                                         'disabled': True}))
 
-    pub_date = SelectField('Year of publication',
-                           choices=[(str(x), str(x))
-                                    for x in
-                                    range(1970,
-                                          datetime.now().year + 1)],
-                           validators=[check_pub_date],
-                           render_kw=({
-                               'class': 'custom-select mb-2 mr-sm-2 mb-sm-0',
-                               'id': 'mySelect',
-                               'placeholder': 'Year of publication',
-                               'disabled': True}))
-
-    """Authors"""
+    # Authors
     first_name = StringField('First name',
                              validators=[check_author],
                              render_kw=({'class': 'inputs',
@@ -154,12 +152,25 @@ class BookForm(FlaskForm):
                                         'placeholder': 'Surname 2',
                                         'disabled': True}))
 
+
+class MagazineForm(LibraryItemForm):
+    title_of_magazine = StringField('Title',
+                                    validators=[DataRequired(),
+                                                Length(3)],
+                                    render_kw=({'class': 'inputs',
+                                                'id': 'title_of_magazine',
+                                                'placeholder':
+                                                    'Title of magazine',
+                                                'disabled': True}))
+
     issue = StringField('Issue',
                         render_kw=({'class': 'inputs',
                                     'id': 'issue',
                                     'placeholder': 'Issue',
                                     'disabled': True}))
 
+
+class MixedForm(MagazineForm, BookForm):
     submit = SubmitField('Create',
                          render_kw=({'class': 'btn btn-primary submits',
                                      'id': 'button',