+++ models/books.py
@@ -3,6 +3,14 @@ from sqlalchemy.orm import class_mapper, ColumnProperty
 from init_db import db
 from models.library import LibraryItem
 
+book_author = db.Table('books_authors',
+                       db.Column('author_id',
+                                 db.Integer,
+                                 db.ForeignKey('authors.id')),
+                       db.Column('book_id',
+                                 db.Integer,
+                                 db.ForeignKey('books.id')))
+
 
 class Book(LibraryItem):
     __tablename__ = "books"
@@ -64,10 +72,3 @@ class Author(db.Model):
 
     def __repr__(self):
         return "<Author: {} {}>".format(self.first_name, self.last_name)
-
-
-book_author = db.Table(
-    "books_authors",
-    db.Column("author_id", db.Integer, db.ForeignKey("authors.id")),
-    db.Column("book_id", db.Integer, db.ForeignKey("books.id")),
-)